refactor: remove CASL module and related dependencies; implement global validation pipe and update user permissions

+++ src/modules/comments/comments.controller.ts
@@ -1,11 +1,20 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, Query } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Patch,
+  Param,
+  Delete,
+  Query,
+} from '@nestjs/common';
 import { CommentsService } from './comments.service';
 import { CreateCommentDto } from './dto/create-comment.dto';
 import { UpdateCommentDto } from './dto/update-comment.dto';
 import { PaginationDto } from 'src/common/dto/pagination.dto';
 import { CurrentUser } from 'src/common/decorators/user.decorator';
 import { JwtPayload } from 'src/common/guards/jwt-auth.guard';
-import { ApiBearerAuth, ApiOperation } from '@nestjs/swagger';
+import { ApiBearerAuth, ApiOperation, ApiQuery } from '@nestjs/swagger';
 
 @ApiBearerAuth('access-token')
 @Controller('comments')
@@ -17,7 +26,10 @@ export class CommentsController {
     description: 'Tạo comment mới cho một bài viết.',
   })
   @Post()
-  create(@CurrentUser() user: JwtPayload, @Body() createCommentDto: CreateCommentDto) {
+  create(
+    @CurrentUser() user: JwtPayload,
+    @Body() createCommentDto: CreateCommentDto,
+  ) {
     return this.commentsService.create(user.sub, createCommentDto);
   }
 
@@ -25,12 +37,11 @@ export class CommentsController {
     summary: 'Danh sách comment của bài viết',
     description: 'Lấy danh sách comment của một bài viết (có phân trang).',
   })
+  @ApiQuery({ name: 'cursor', required: false, example: 1 })
+  @ApiQuery({ name: 'take', required: false, example: 10 })
   @Get()
-  findComments(
-    @Param('postId') postId: string,
-    @Query() query: PaginationDto,
-  ) {
-    return this.commentsService.findComments(postId, query)
+  findComments(@Param('postId') postId: string, @Query() query: PaginationDto) {
+    return this.commentsService.findComments(postId, query);
   }
 
   @ApiOperation({
@@ -38,8 +49,16 @@ export class CommentsController {
     description: 'Chỉ chủ comment mới được phép cập nhật.',
   })
   @Patch(':id')
-  update(@Param('id') id: string, @CurrentUser() user: JwtPayload, @Body() updateCommentDto: UpdateCommentDto) {
-    return this.commentsService.update(id, user.sub, updateCommentDto);
+  update(
+    @Param('id') id: string,
+    @CurrentUser() user: JwtPayload,
+    @Body() updateCommentDto: UpdateCommentDto,
+  ) {
+    return this.commentsService.update(
+      id,
+      { sub: user.sub, role: user.role },
+      updateCommentDto,
+    );
   }
 
   @ApiOperation({
@@ -47,7 +66,7 @@ export class CommentsController {
     description: 'Chỉ chủ comment mới được phép xoá.',
   })
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.commentsService.remove(id);
+  remove(@Param('id') id: string, @CurrentUser() user: JwtPayload) {
+    return this.commentsService.remove(id, { sub: user.sub, role: user.role });
   }
 }
